Move `init_rollup_store` to `l2::initializers`
